Relationship schedule with Profitability-Structuring Results and fix details
Connextion OnetoOne with the tables and fix names and date tipe in someone values

+++ src/main/java/com/bonces/webservicesbonces/results/domain/model/entity/StructuringResults.java
@@ -39,22 +39,22 @@ public class StructuringResults  extends AuditModel {
 
 
     @NotNull
-    private long effectiveAnnualRate;
+    private double effectiveAnnualRate;
 
 
     @NotNull
-    private long effectiveRate;
+    private double effectiveRate;
 
 
     @NotNull
-    private long COK;
+    private double COK;
 
 
     @NotNull
-    private long InitialCostsEmitter;
+    private double InitialCostsEmitter;
 
     @NotNull
-    private long InitialCostsBondholder;
+    private double InitialCostsBondholder;
 
     @OneToOne(fetch = FetchType.LAZY)
     @JoinColumn(
